Update LLVM commit to 829616c (#814)
* Change LLVM::DialectCastOp to UnrealizedConversionCastOp.

Reflected a change in https://reviews.llvm.org/D105880

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Change llvmVoidTy to TypeRange() to meet result type verification in llvm.call

Reflected a change in https://reviews.llvm.org/D106937

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Update MLIR comit to b44eb5a in README.md and a script.

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Remove unused variable.

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Update MLIR comit to 12b34e0 in README.md and a script.

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Use math-to-llvm

Reflected a changes in https://reviews.llvm.org/D105702 about factor
math-to-llvm out of standard-to-llvm.

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Fixed type mismatch in LLVM::InsertValueOp

Got type mismatch error after added verifier
in https://reviews.llvm.org/D106398. This commit fixes it.

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Move SimplifyKrnl.cpp to src/Dialect/Krnl/ to avoid cyclic dependency.

After a patch D105903 was merged in MLIR, got link-time failure in shared
library build. This is because the patch forces the compiler to emit a
strong definition for a class. To avoid this, OMKrnlOps need to be added
in LINK_LIBS of OMSimplifyKrnl. However, this introduces cyclic
dependency(OMSimplifyKrnl -> OMKrnlOps -> OMSimplifyKrnl). So, this
commit moves SimplifyKrnl.cpp in OMKrnlOps to avoid cyclic dependency.

Reference: https://reviews.llvm.org/D105903

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Removed src/Transform/Krnl in src/Transform/CMakeLists.txt

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Update MLIR comit to ab03a84 in README.md and a script

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Update MLIR comit to c84c5d0 in README.md and a script

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Updated MLIR commit to 0104cc

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

* Updated MLIR commit to 829616c

Signed-off-by: Haruki Imai <imaihal@jp.ibm.com>

Co-authored-by: Alexandre Eichenberger <alexe@us.ibm.com>