Run verification code before shape inference (#1358)
Currently ShapeInferencePass calls inferShapes for each operation in a MLIR Region. This causes inferShapes to be run on an operation after has been run on all of its operands. It is assumed that the operation is legal once its inferShapes function is run, however this is not the case for operations that had operands with unknown shape and rank when the verification code was first executed (the verification code was skipped in that situation).

This PR causes the verification code to be called on an operation immediately before its inferShape function is called. This guarantees that the proper diagnostics are emitted for operations that had operands with unknown shape when the verification code was first executed.

Signed-off-by: Ettore Tiotto <etiotto@ca.ibm.com>