Verify the input tensors of the entry point function (#1472)
This patch adds verification of the input tensors that are passed to entry point functions such as run_main_graph. Entry point signature is used to verify tensors. In particular, it verifies:
- The number of input tensors
- Data type of each input tensor
- Rank and shape of each input tensor

When the input tensors are invalid, run_main_graph returns NULL and errno = EINVAL (Invalid argument) is emitted.

A new compile option --verifyInputTensors is added to enable verification inside the entry point function. It's off by default.

A new backend test is added to verify this verification, that is make check-onnx-backend-input-verification.

Signed-off-by: Tung D. Le <tung@jp.ibm.com>