Fix memory leak due to unallocated tensor data (#2176)
Fixes https://github.com/onnx/onnx-mlir/issues/2171. The Java `OMTensor` object was not taking full ownership duties of the backing data within the native `OMTensor` object, leading to a memory leak. This fix solves the issue by allocating a new `ByteBuffer` in the JNI `OMTensor` constructor with a copy, leaving the original data to be freed along with the jni `OMTensor`.

Some initial timings suggest that there isn't any meaningful difference in timing due to the copy in the YOLOv3-12 and MNIST models.