Update CPU and NNPA docs with min and max supported opsets  (#2452)
# Description
* Refactor documentation to show supported model opset range for each individual operator.
* Add DeformConv to unsupported operator list (looks like we missed it on going to Opset 19)

# "Summary" / Novella
There are now two "views" of Op versions. The "code" view (ie .py and CMakeLists.txt) uses `==Min==` and `==UNSUPPORTED==` values. They are set based on the versions of the operator we know work (ie an operator centric view). 

The "documentation" view (ie *.md files) is Opset centric. The values presented there are based on model opsets and attempt to mask the complexity of how ONNX versions individual Operators. (ie Opset 19 uses the last updated version of Add, which is currently 14. So while it is technically accurate to say we only support "up to" Add 14, users may interpret that to mean their model's Opset must be less then or equal to <= 14, which isn't true.

For this PR, I set the Ops' `==MIN==` in the code files based on passing models from the onnx-model zoo. For example say a model is based on Opset 6, has Operator X, compiles, and successfully runs with onnx-mlir. Further say Operator X has versions 1, 2, 7, 10 and 19. Since the Opset 6 model works and uses X, then what we have in onnx-mlir is likely minimally compatible with X version 2 (ie 2 is the latest version of X that's <= the model's opset 6). This is not a perfect test but it is a good launching point and lines up with the models we show as supported in our model-zoo testing.

In the code view, `==UNSUPPORTED==` (if set), indicates an Op we support an earlier version of, but have not updated for a particular Opset. For example, if supported Op X was updated in Opset 19, but onnx-mlir has not been updated to handle X 19, then X's `==UNSUPORTED==` will be set to 19. If we support the latest version of the Op available, `==UNSUPPORTED==` is not set.

Since users and models are generally Opset focused, we make our documentation Opset centric. As a general statement, we only support models with a minimum Opset of 6. So continuing the example of Operator X, it has a `==MIN==` of 2 and `==UNSUPPORTED` of 19. Then our documentation (ie support statement) is that we only support models that contain Operator X if the model's Opset is between 6 and 18 (inclusvie). 

The documenation's Opset minimum for each Op is calculated as `max( Op's ==MIN==, 6)`. The Opset maximum for each Op is the Op's` ==UNSUPPORTED==` - 1 or `*`. Where `*` indicates we support the latest version available at onnx-mlir release. (The latest Opset available at release is specified at the top of the document view).